fix(audit): pass 3 — real collapse fix, non-vacuous size-cap test, overwrite naming
Pass 2's save-card collapse fix did NOT work. Proven with qml6: a
self-assignment does not drop a QML binding when the value is unchanged, so
the binding stayed live and the card would still have slammed shut on the
user's first save. SettingsCard now owns the semantics with an
`initiallyCollapsed` property, read once at construction and applied without
the collapse animation. Verified with qml6 on all four counts: starts closed,
no flash-then-collapse, a later save does not re-collapse, and a card the user
expanded by hand stays where they put it.

Pass 2's oversized-import test was VACUOUS: 5 MiB of 'x' is rejected by the
JSON parser, not the size cap, so it passed with the cap deleted. It now pads
a VALID payload past the cap. The chmod-000 snapshot test now skips under root
(the project's documented Docker flow runs as root, where permissions are not
enforced and the test would fail).

Store: removeSet's missing-file path was still silent (pass 2 only fixed the
could-not-delete branch), and the empty-snapshot refusal never said why. Both
toast now, so every failure carries its reason and QML no longer infers one
from a bare false (the contradictory save banner is gone). Every
pendingChangesChanged goes through the one gate. setExists became
existingSetName: names collide by SLUG, so the confirm prompt now names the
STORED set rather than the user's spelling of it.

AnimationPresetLibrary was the last consumer discarding snapshotFileIfFirst's
failure bool, and so the last unguarded data-loss path: it now refuses the
write like every other caller. Adopted MetadataChip in ShaderBrowserCard (the
one other place the badge recipe was copy-pasted), decoded percent-encoded
basenames in the install banner, and made the two toast call sites use the
KCM-safe window guard the rest of the codebase uses.

Tests: split the user-preset half into test_animations_presets.cpp — the
motion file had grown to 825 lines against the 800-line cap while its own
header claimed the split existed to stay under it. Hoisted the duplicated
TreeStubSettings and rowFor helpers into tests/unit/helpers. The coalesced-
refresh test now asserts its drain actually delivered, the sets-dir teardown
refuses to run outside QStandardPaths test mode, and the mutation guard covers
all five mutators.

Fixed the last stale CMake counts: three ShaderBrowserPage wrappers (not two),
four AnimationsPageController detail-header consumers (not two), and one
DecorationPageController consumer (not two).
