+++ tests/unit/settings/test_animations_motion_sets.cpp
@@ -3,17 +3,17 @@
 
 /**
  * @file test_animations_motion_sets.cpp
- * @brief AnimationsPageController user-preset, motion-set, and pending-changes tests.
+ * @brief AnimationsPageController motion-set and pending-changes tests.
  *
- * Split from test_animations_page_controller.cpp to keep each test file
- * under the project's <800-line guideline.
+ * Split from test_animations_page_controller.cpp, and the user-preset half
+ * split again into test_animations_presets.cpp, to keep each test file under
+ * the project's <800-line guideline.
  *
  * Motion sets go through the shared ShaderSetStore reached from
  * `AnimationsPageController::setsBridge()`. The domain half (which paths are
  * valid, how live state is snapshotted) lives in motionsetdomain.cpp; the
  * envelope, the file lifecycle and the `active` summary are the store's.
  * Pinned behaviour:
- *   - User preset CRUD (addUserPreset, userPresets, removeUserPreset)
  *   - Motion set save/apply/remove with merge-not-replace semantics
  *   - `active` is a CONTAINMENT check: a set stays active while unrelated
  *     overrides exist, because apply would have left them alone
@@ -28,8 +28,6 @@
  *   - The in-flight-discard mutation guard refuses every set write
  *   - Pending changes signal emission for revert/commit
  *   - Atomic motion-set application (rejects whole malformed set)
- *   - Cross-cutting safety: removeUserPreset must not touch override files,
- *     malformed preset JSON logs+skips
  */
 
 #include <QSignalSpy>
@@ -48,26 +46,11 @@
 
 #include "settings/animationspagecontroller.h"
 #include "settings/shadersetstore.h"
+#include "../helpers/SetRowHelpers.h"
+
+#include <unistd.h>
 
 using namespace PlasmaZones;
-
-namespace {
-
-/// The row for @p name, or an empty map when no such set is listed. Callers
-/// QVERIFY the result is non-empty BEFORE asserting on a field, so a missing
-/// row fails loudly instead of skipping the assertion.
-QVariantMap rowFor(ShaderSetStore* sets, const QString& name)
-{
-    const QVariantList rows = sets->availableSets();
-    for (const QVariant& row : rows) {
-        const QVariantMap map = row.toMap();
-        if (map.value(QStringLiteral("name")).toString() == name)
-            return map;
-    }
-    return QVariantMap{};
-}
-
-} // namespace
 
 class TestAnimationsMotionSets : public QObject
 {
@@ -81,180 +64,6 @@ private Q_SLOTS:
     void initTestCase()
     {
         QStandardPaths::setTestModeEnabled(true);
-    }
-
-    // ─── User preset library ──────────────────────────────────────────────
-
-    void addUserPreset_writesFileWithSlugFilename()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        QSignalSpy spy(&c, &AnimationsPageController::userPresetsChanged);
-        QVERIFY(c.addUserPreset(
-            QStringLiteral("My Snappy Spring!"),
-            {{QStringLiteral("curve"), QStringLiteral("spring:14.0,0.7")}, {QStringLiteral("duration"), 200}}));
-        QCOMPARE(spy.count(), 1);
-
-        // Slugified filename: lowercase, non-alnum collapsed to '-',
-        // trailing '-' trimmed.
-        QVERIFY(QFileInfo::exists(tmp.path() + QStringLiteral("/my-snappy-spring.json")));
-    }
-
-    void userPresets_excludesPathNamedFiles()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        // Preset file
-        QVERIFY(
-            c.addUserPreset(QStringLiteral("My Curve"), {{QStringLiteral("curve"), QStringLiteral("0.5,0,0.5,1")}}));
-        // Override file at a known path
-        QVERIFY(c.setOverride(QStringLiteral("editor.snapIn"), {{QStringLiteral("duration"), 250}}));
-
-        // userPresets sees ONLY the preset, not the override
-        const QVariantList presets = c.userPresets();
-        QCOMPARE(presets.size(), 1);
-        QCOMPARE(presets.first().toMap().value(QStringLiteral("name")).toString(), QStringLiteral("My Curve"));
-    }
-
-    /// An orphan override file at a path that this build no longer
-    /// recognises (e.g. left over after a taxonomy rename like PR #400's
-    /// `panel.popup.*` → `popup.*`) MUST NOT leak into the preset list.
-    /// `userPresets` previously filtered only on the current
-    /// `allBuiltInPaths()`; an orphan file at the obsolete path would
-    /// pass that gate and surface as a fake preset named after the
-    /// obsolete event path. The guard is the basename-contains-dot
-    /// check: `setOverride` writes verbatim path filenames (always
-    /// dotted for non-root paths), `addUserPreset` slugifies (strips
-    /// dots), so a dotted basename is the override-file fingerprint.
-    void userPresets_excludesOrphanOverrideFiles()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        // Plant an orphan override file directly: the on-disk shape
-        // setOverride would have written for the obsolete path
-        // `panel.popup` (which existed pre-PR-400 but isn't in
-        // `allBuiltInPaths()` post-rename). `setOverride` itself rejects
-        // unknown paths now, so we hand-craft the file the way an older
-        // build would have.
-        QFile f(tmp.path() + QStringLiteral("/panel.popup.json"));
-        QVERIFY(f.open(QIODevice::WriteOnly));
-        QJsonObject obj;
-        obj.insert(QStringLiteral("name"), QStringLiteral("panel.popup"));
-        obj.insert(QStringLiteral("duration"), 1000);
-        obj.insert(QStringLiteral("curve"), QStringLiteral("0.33,1,0.68,1"));
-        const QByteArray objBytes = QJsonDocument(obj).toJson();
-        QCOMPARE(f.write(objBytes), static_cast<qint64>(objBytes.size()));
-        f.close();
-
-        // Add a real preset alongside.
-        QVERIFY(c.addUserPreset(QStringLiteral("My Curve"), {{QStringLiteral("duration"), 200}}));
-
-        // userPresets must see ONLY the real preset; the orphan file
-        // is a stale override, not a preset.
-        const QVariantList presets = c.userPresets();
-        QCOMPARE(presets.size(), 1);
-        QCOMPARE(presets.first().toMap().value(QStringLiteral("name")).toString(), QStringLiteral("My Curve"));
-    }
-
-    void addUserPreset_rejectsKnownPathNames()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        QSignalSpy spy(&c, &AnimationsPageController::userPresetsChanged);
-        // "editor.snapIn" is a known event path — would shadow the
-        // override slot.
-        QVERIFY(!c.addUserPreset(QStringLiteral("editor.snapIn"), {{QStringLiteral("duration"), 100}}));
-        QCOMPARE(spy.count(), 0);
-        QVERIFY(c.userPresets().isEmpty());
-    }
-
-    void addUserPreset_rejectsEmptyAndAllSymbol()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        QVERIFY(!c.addUserPreset(QString(), {{QStringLiteral("duration"), 100}}));
-        // All-symbol slugifies to empty → reject (would write to ".json")
-        QVERIFY(!c.addUserPreset(QStringLiteral("@@@"), {{QStringLiteral("duration"), 100}}));
-    }
-
-    void removeUserPreset_emitsAndDeletes()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        QVERIFY(
-            c.addUserPreset(QStringLiteral("To Delete"), {{QStringLiteral("curve"), QStringLiteral("spring:10,0.5")}}));
-        QCOMPARE(c.userPresets().size(), 1);
-
-        QSignalSpy spy(&c, &AnimationsPageController::userPresetsChanged);
-        QVERIFY(c.removeUserPreset(QStringLiteral("To Delete")));
-        QCOMPARE(spy.count(), 1);
-        QVERIFY(c.userPresets().isEmpty());
-    }
-
-    void removeUserPreset_unknownReturnsFalse()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        QSignalSpy spy(&c, &AnimationsPageController::userPresetsChanged);
-        QVERIFY(!c.removeUserPreset(QStringLiteral("nonexistent")));
-        QCOMPARE(spy.count(), 0);
-    }
-
-    /// `removeUserPreset` MUST not delete an override file even when its
-    /// embedded `name` field happens to match the supplied preset name.
-    /// Pre-fix, the directory-scan fallback walked every JSON file in
-    /// the profiles dir and matched by `name`; an override file whose
-    /// `name` matched the searched preset would be deleted.
-    void removeUserPreset_doesNotTouchOverrideFiles()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        // Write an override file. Its on-disk `name` field is "editor.snapIn"
-        // (per setOverride's stamping rule).
-        QVERIFY(c.setOverride(QStringLiteral("editor.snapIn"), {{QStringLiteral("duration"), 250}}));
-        const QString overrideFilePath = tmp.path() + QStringLiteral("/editor.snapIn.json");
-        QVERIFY(QFileInfo::exists(overrideFilePath));
-
-        // Hand-craft a hostile preset on disk: `name` field = "editor.snapIn"
-        // but the FILE is not at the canonical override slot. The library
-        // skips it on userPresets() (collision filter) but a naive
-        // remove-by-name walk would still delete it. We're asserting the
-        // override file in the canonical slot remains intact regardless.
-        QSignalSpy spy(&c, &AnimationsPageController::userPresetsChanged);
-        QVERIFY(!c.removeUserPreset(QStringLiteral("editor.snapIn")));
-        QCOMPARE(spy.count(), 0);
-
-        // The override file MUST still exist — this is the load-bearing
-        // assertion: removeUserPreset cannot collateral-damage overrides.
-        QVERIFY2(
-            QFileInfo::exists(overrideFilePath),
-            "override file was deleted by removeUserPreset(\"editor.snapIn\") — preset CRUD MUST NOT touch override "
-            "slots");
-        QCOMPARE(c.rawProfile(QStringLiteral("editor.snapIn")).value(QStringLiteral("duration")).toInt(), 250);
     }
 
     // ─── Motion sets ──────────────────────────────────────────────────────
@@ -501,40 +310,6 @@ private Q_SLOTS:
         QCOMPARE(overrideSpy.count(), 0);
     }
 
-    // ─── Logging on malformed JSON ────────────────────────────────────────
-
-    /// Plant an unparseable JSON file in the profiles dir; userPresets()
-    /// should skip it AND log a warning at the qCWarning level. The
-    /// emission is the load-bearing piece — pre-fix the parse error was
-    /// silently swallowed.
-    void userPresets_malformedJsonLogsAndSkips()
-    {
-        QTemporaryDir tmp;
-        QVERIFY(tmp.isValid());
-        AnimationsPageController c;
-        c.setUserProfilesDirOverride(tmp.path());
-
-        // Write a known-good preset so the iteration has at least one
-        // success result.
-        QVERIFY(c.addUserPreset(QStringLiteral("Good"), {{QStringLiteral("duration"), 100}}));
-
-        // Plant a malformed file directly.
-        QFile bad(tmp.path() + QStringLiteral("/garbage.json"));
-        QVERIFY(bad.open(QIODevice::WriteOnly));
-        bad.write("{ this is not valid json");
-        bad.close();
-
-        // Expect a warning to be logged for the malformed file. The
-        // exact message is "AnimationPresetLibrary: failed to parse <path> : <error>".
-        QTest::ignoreMessage(QtWarningMsg,
-                             QRegularExpression(QStringLiteral("AnimationPresetLibrary: failed to parse.*garbage")));
-
-        const QVariantList presets = c.userPresets();
-        // The good preset still surfaces; the malformed one is skipped.
-        QCOMPARE(presets.size(), 1);
-        QCOMPARE(presets.first().toMap().value(QStringLiteral("name")).toString(), QStringLiteral("Good"));
-    }
-
     // ─── Motion sets: active flag, metadata edit, portability, guard ───────
 
     /// `active` measures the saved payload against the CURRENT override files.
@@ -618,7 +393,7 @@ private Q_SLOTS:
 
         QVERIFY(c.setOverride(QStringLiteral("editor.snapIn"), {{QStringLiteral("duration"), 250}}));
         QVERIFY(sets->saveCurrentAsSet(QStringLiteral("Taken"), QStringLiteral("first")));
-        QVERIFY(sets->setExists(QStringLiteral("Taken")));
+        QCOMPARE(sets->existingSetName(QStringLiteral("Taken")), QStringLiteral("Taken"));
 
         QVERIFY(c.setOverride(QStringLiteral("editor.snapIn"), {{QStringLiteral("duration"), 999}}));
         QSignalSpy toastSpy(sets, &ShaderSetStore::toastRequested);
@@ -717,6 +492,13 @@ private Q_SLOTS:
     /// destroying content Discard could not restore.
     void motionSets_refusesWriteWhenSnapshotFails()
     {
+        // The guard is provoked by making an existing set file unreadable, and
+        // root bypasses file permissions entirely — the read would succeed and
+        // the guard would never fire. Skip rather than fail in the project's
+        // documented Docker flow, which runs as root.
+        if (::geteuid() == 0)
+            QSKIP("chmod 000 is not enforced for root");
+
         QTemporaryDir tmp;
         QVERIFY(tmp.isValid());
         AnimationsPageController c;
@@ -799,6 +581,12 @@ private Q_SLOTS:
         QVERIFY(c.setOverride(QStringLiteral("editor.snapIn"), {{QStringLiteral("duration"), 250}}));
         QVERIFY(sets->saveCurrentAsSet(QStringLiteral("Before"), QString()));
 
+        // A file on hand for the importSet refusal below.
+        QTemporaryDir exportDir;
+        QVERIFY(exportDir.isValid());
+        const QString exported = exportDir.filePath(QStringLiteral("before.json"));
+        QVERIFY(sets->exportSet(QStringLiteral("Before"), exported));
+
         // Kick off the async discard and immediately try to write, without
         // spinning the event loop — the worker is in flight.
         c.asyncRevertPending();
@@ -810,14 +598,16 @@ private Q_SLOTS:
         QVERIFY2(!sets->removeSet(QStringLiteral("Before")), "a remove must be refused mid-discard");
         QVERIFY2(!sets->updateSet(QStringLiteral("Before"), QStringLiteral("Renamed"), QString()),
                  "an update must be refused mid-discard");
-        QCOMPARE(toastSpy.count(), 4); // each refusal carries its reason
+        QVERIFY2(!sets->importSet(exported), "an import must be refused mid-discard");
+        QCOMPARE(toastSpy.count(), 5); // each refusal carries its reason
+
+        // Let the worker finish before reading the dir back — it is rewriting
+        // the same tree underneath us.
+        QSignalSpy done(&c, &PhosphorControl::StagingDomain::discardResult);
+        QVERIFY(done.wait(5000));
 
         // The refused save never reached disk.
         QVERIFY(rowFor(sets, QStringLiteral("During")).isEmpty());
-
-        // Let the worker finish so the fixture tears down cleanly.
-        QSignalSpy done(&c, &PhosphorControl::StagingDomain::discardResult);
-        QVERIFY(done.wait(5000));
     }
 };
